Remove unused imports, fix coding style
